Add pool manager and factory
