+++ package.json
@@ -18,6 +18,7 @@
     "dev:landing": "scripts/local/dev -a \"landing\" -m \"$npm_config_mock\"",
     "lint": "eslint --ext .vue,.ts ./ --fix",
     "test": "echo \"Error: no test specified\" && exit 1",
+    "test:ethereum": "scripts/ethereum/test -f \"$npm_config_fork\"",
     "test:push": "scripts/actions/test -w push",
     "release": "release-it",
     "deploy:sample": "npm run deploy:sample --workspace @casimir/evm",
